moved several classes & enums to separate files
