+++ Data/UniqueBindDirection.cs
@@ -1,0 +1,9 @@
+﻿namespace Needleforge.Data;
+
+public enum UniqueBindDirection
+{
+    UP,
+    DOWN, 
+    LEFT, 
+    RIGHT
+}
