+++ Data/VanillaCrest.cs
@@ -1,0 +1,16 @@
+﻿namespace Needleforge.Data;
+
+public enum VanillaCrest
+{
+    HUNTER,
+    HUNTER_V2,
+    HUNTER_V3,
+    BEAST,
+    REAPER,
+    WANDERER,
+    WITCH,
+    ARCHITECT,
+    SHAMAN,
+    CURSED,
+    CLOAKLESS,
+}
